code refactoring: updated variables and file names
